+++ src/main.js
@@ -1,15 +1,13 @@
 import {Menu} from './components/menu';
 import {Filters} from './components/filters';
-import {EventEdit} from './components/trip-event-edit-form';
 import {TripInfo} from './components/trip-info';
-import {TripDay} from './components/trip-day';
 import {TripDayList} from './components/trip-day-list';
-import {EventList, EmptyEventList} from './components/trip-event-list';
-import {Event} from './components/trip-event';
+import {EmptyEventList} from './components/trip-event-list';
 import {getEvent, getFilters} from './data';
 import {render, Position} from './utils';
+import {TripController} from './components/trip-controller';
 
-const EVENT_COUNT = 0;
+const EVENT_COUNT = 6;
 
 const events = Array(EVENT_COUNT).fill().map(()=> getEvent())
   .sort((event1, event2)=> event1.timeStart - event2.timeStart);
@@ -34,54 +32,9 @@ events.map(((event)=>{
 tripInfoCostElement.querySelector(`.trip-info__cost-value`).textContent = totalAmount;
 
 const tripEventsElement = document.querySelector(`.trip-events`);
-const tripDaysElement = events.length > 0 ? new TripDayList().getElement() : new EmptyEventList().getElement();
-render(tripEventsElement, tripDaysElement);
-
-const days = new Set(events.map(({timeStart})=>(new Date(timeStart)).setHours(0, 0, 0, 0)));
-
-Array.from(days).forEach((day, index) => {
-  let dayElement = new TripDay(day, index + 1).getElement();
-  render(tripDaysElement, dayElement);
-
-  const eventsList = new EventList().getElement();
-  render(dayElement, eventsList);
-
-  events.filter(({timeStart}) => new Date(day).toLocaleDateString() === new Date(timeStart).toLocaleDateString()).forEach((event)=> {
-    let eventComponent = new Event(event);
-    let eventEditComponent = new EventEdit(event);
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === `Escape` || evt.key === `Esc`) {
-        eventsList.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    eventComponent.getElement()
-      .querySelector(`.event__rollup-btn`)
-      .addEventListener(`click`, () => {
-        eventsList.replaceChild(eventEditComponent.getElement(), eventComponent.getElement());
-        document.addEventListener(`keydown`, onEscKeyDown);
-      });
-
-    eventEditComponent.getElement()
-      .querySelector(`.event__rollup-btn`)
-      .addEventListener(`click`, () => {
-        eventsList.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
-        document.addEventListener(`keydown`, onEscKeyDown);
-      });
-
-    eventEditComponent.getElement()
-      .querySelector(`.event__save-btn`)
-      .addEventListener(`click`, () => {
-        eventsList.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      });
 
 
-    render(eventsList, eventComponent.getElement());
-  });
-});
+new TripController(events, tripEventsElement).init();
 
 const tripControlsElement = document.querySelector(`.trip-main__trip-controls`);
 render(tripControlsElement, new Filters(getFilters(events)).getElement());
